Update revealjs-url
lab.hakim.se/reveal-js redirects to revealjs.com, but it must be first
doing an http redirect, because trying to load a presentation (generated
with this plugin) over https fails to correctly fetch the css and
javascript files. The Chrome developer console complains that it is
trying to load "unsafe" scripts from http, even though they're https.

Changing this url should fix that.